Update with JSON parsing experiments
Add map and fixtures for testing json parsing both generically as well
as into a very basic struct.  Long term plan is to use the Struct to
house the user provided column map (with known structure) and the
generic to house the raw data.

We can eventually use this to setup goroutines to de-identify user
defined arrays with some nice concurrency.  I still feel like a noob,
but we're getting there.